+++ monthly-budget/backend/src/controllers/BudgetController.js
@@ -4,7 +4,7 @@ const {
   TransactionService,
   TransactionType,
   TransactionStatus
-} = require("@klutch-card/klutch-js");
+} = require("@klutch-card/klutch-js")
 const httpStatus = require('http-status')
 const Ajv = require("ajv")
 const { upsertBudget, listBudgets, deleteBudget: findAndDeleteBudget } = require("../models/Budget")
@@ -56,20 +56,6 @@ const handlerTransactionsDataPerBudgets = (budgets, transactions) => {
     // result.transactions = trxs
     return result
   })
-}
-
-const addPanelToHomeScreen = async (recipeInstallId) => {
-  try {
-    const panels = await RecipesService.getPanels(undefined)
-    const panelsRecipeInstallId = panels.map(p => p.recipeInstall.id)
-
-    if (panelsRecipeInstallId.includes(recipeInstallId)) return
-
-    console.log(`adding home panel to recipeInstallId \"${recipeInstallId}\"`)
-    await RecipesService.addPanel(recipeInstallId, "/templates/HomePanel.template", { recipeId }, null)
-  } catch (err) {
-    console.log({ err, recipeInstallId })
-  }
 }
 
 const createOrUpdateBudget = async (req, resp) => {
@@ -131,8 +117,6 @@ const getBudgets = async (req, resp) => {
     console.log({ err })
     return resp.status(httpStatus.INTERNAL_SERVER_ERROR).json({ errorMessage: 'fail build authorization token' })
   }
-
-  if (budgets.length == 0) addPanelToHomeScreen(recipeInstallId)
 
   try {
     const now = new Date()
